requirements.txt
pynput
pyscreenshot
Pillow
